added fuctions to interface actor.go
